fixed errors in time_tools crate
