+++ module/core/time_tools/src/lib.rs
@@ -48,6 +48,10 @@
 /// Collection of time tools.
 // pub mod time;
 
+#[ cfg( feature = "time_now" ) ]
+#[ path = "./now.rs" ]
+pub mod now;
+
 /// Dependencies.
 pub mod dependency
 {
@@ -75,76 +79,12 @@ pub mod exposed
 {
   #[ doc( inline ) ]
   pub use super::prelude::*;
+  #[ cfg( feature = "time_now" ) ]
+  #[ doc( inline ) ]
+  pub use super::now::*;
 }
 
 /// Prelude to use essentials: `use my_module::prelude::*`.
 pub mod prelude
 {
-}
-
-use std::time;
-
-///
-/// Get current time. Units are milliseconds.
-///
-
-pub fn now() -> i64
-{
-  time::SystemTime::now()
-  .duration_since( time::UNIX_EPOCH ).unwrap()
-  .as_millis() as i64
-}
-
-///
-/// Default units are seconds.
-///
-
-pub mod s
-{
-  use super::*;
-
-  /// Get current time. Units are seconds.
-  pub fn now() -> i64
-  {
-    time::SystemTime::now()
-    .duration_since( time::UNIX_EPOCH ).unwrap()
-    .as_secs() as i64
-  }
-}
-
-///
-/// Default units are milliseconds.
-///
-
-pub mod ms
-{
-  use super::*;
-
-  /// Get current time. Units are milliseconds.
-  pub fn now() -> i64
-  {
-    time::SystemTime::now()
-    .duration_since( time::UNIX_EPOCH ).unwrap()
-    .as_millis() as i64
-  }
-}
-
-// xxx : qqq for Dima : problem. ms should not be part of `wtools::ms`, something is wrong. fix it, please
-/* aaa : Dmytro : all routines and modules is inside wtools and wtools::time, added test suite to test it */
-
-///
-/// Default units are nanoseconds.
-///
-
-pub mod ns
-{
-  use super::*;
-
-  /// Get current time. Units are nanoseconds.
-  pub fn now() -> i64
-  {
-    time::SystemTime::now()
-    .duration_since( time::UNIX_EPOCH ).unwrap()
-    .as_nanos() as i64
-  }
 }
