--- module/core/time_tools/tests/inc/inc.rs
@@ -1,5 +1,0 @@
-#[ allow( unused_imports ) ]
-use super::*;
-#[ allow( unused_imports ) ]
-use test_tools::exposed::*;
-
