+++ module/core/time_tools/tests/time_tests.rs
@@ -1,5 +1,6 @@
 
 #[ allow( unused_imports ) ]
 use test_tools::exposed::*;
+use time_tools as TheModule;
 
 mod inc;
